// web/app/domain/d7d/types.ts
export type TDice = "d4" | "d6" | "d8" | "d10" | "d12" | "d20" | "d100";

export const enum ESkill {
    Acrobatics = 1,
    AnimalHandling = 2,
    Arcana = 3,
    Athletics = 4,
    Deception = 5,
    History = 6,
    Insight = 7,
    Intimidation = 8,
    Investigation = 9,
    Medicine = 10,
    Nature = 11,
    Perception = 12,
    Performance = 13,
    Persuasion = 14,
    Religion = 15,
    SleightOfHand = 16,
    Stealth = 17,
    Survival = 18,
}